Сделал анимацию

+++ app/src/main/java/com/ermilov/focushomeworktwo/SpeedChangeListener.kt
@@ -1,5 +1,5 @@
 package com.ermilov.focushomeworktwo
 
 interface SpeedChangeListener {
-    fun onSpeedChanged(newSpeedValue: Float)
+    fun onSpeedChanged(newSpeedValue: Int)
 }
